+++ ext/agoo/page.c
@@ -384,7 +384,7 @@ page_immutable(Err err, const char *path, const char *content, int clen) {
 	p->path = NULL;
     } else {
 	p->path = strdup(path);
-	plen = strlen(path);
+	plen = (int)strlen(path);
 	DEBUG_ALLOC(mem_page_path, p->path);
     }
     p->mtime = 0;
@@ -510,7 +510,7 @@ update_contents(Page p) {
 
 static void
 page_remove(Page p) {
-    int		len = strlen(p->path);
+    int		len = (int)strlen(p->path);
     int64_t	h = calc_hash(p->path, &len);
     Slot	*bucket = get_bucketp(h);
     Slot	s;
@@ -677,7 +677,7 @@ group_create(const char *path) {
 	g->next = cache.groups;
 	cache.groups = g;
 	g->path = strdup(path);
-	g->plen = strlen(path);
+	g->plen = (int)strlen(path);
 	DEBUG_ALLOC(mem_group_path, g->path);
 	g->dirs = NULL;
     }
@@ -693,7 +693,7 @@ group_add(Group g, const char *dir) {
 	d->next = g->dirs;
 	g->dirs = d;
 	d->path = strdup(dir);
-	d->plen = strlen(dir);
+	d->plen = (int)strlen(dir);
 	DEBUG_ALLOC(mem_dir_path, d->path);
     }
 }
